Refactor phdc endpoint to include report_type (#1147)
Endpoint report type
