+++ containers/message-parser/tests/test_parse_message_endpoint.py
@@ -1,4 +1,5 @@
 import json
+from copy import deepcopy
 from pathlib import Path
 from unittest import mock
 
@@ -387,4 +388,46 @@ def test_parse_message_neither_internal_nor_external_schema():
         actual_response.json()["detail"][0]["msg"]
         == "Values for 'parsing_schema' and 'parsing_schema_name' have not been "
         "provided. One, but not both, of these values is required."
+    )
+
+
+def test_schema_without_reference_lookup():
+    no_lookup_schema = deepcopy(test_reference_schema)
+    del no_lookup_schema["labs"]["secondary_schema"]["ordering_provider"][
+        "reference_lookup"
+    ]
+    request = {
+        "message_format": "fhir",
+        "message": {},
+        "parsing_schema": no_lookup_schema,
+    }
+
+    actual_response = client.post("/parse_message", json=request)
+    assert actual_response.status_code == 422
+    assert (
+        actual_response.json()["detail"][0]["msg"]
+        == "Secondary fields in the parsing schema that reference other "
+        "resources must include a `reference_lookup` field that identifies "
+        "where the reference ID can be found."
+    )
+
+
+def test_schema_without_identifier_path():
+    no_bundle_schema = deepcopy(test_reference_schema)
+    no_bundle_schema["labs"]["secondary_schema"]["ordering_provider"][
+        "fhir_path"
+    ] = "Observation.provider"
+    request = {
+        "message_format": "fhir",
+        "message": {},
+        "parsing_schema": no_bundle_schema,
+    }
+
+    actual_response = client.post("/parse_message", json=request)
+    assert actual_response.status_code == 422
+    assert (
+        actual_response.json()["detail"][0]["msg"]
+        == "Secondary fields in the parsing schema that provide `reference_lookup` "
+        "locations must have a `fhir_path` that begins with `Bundle` and identifies "
+        "the type of resource being referenced."
     )
